fix(audit): PR #535 senior-engineer pass — qFatal null guard, dtor disconnect, two-phase host create, test split
Pass 6 of the PR #535 audit, addressing findings from a six-partition
parallel reviewer sweep.

Behaviour fixes
- PopoutController: replace silent release-build crash on null transport
  with qFatal so the precondition failure is diagnosable in both debug
  and release.
- InAppPopoutTransport destructor: disconnect dismissed signal before
  delete so Component.onDestruction's emit cannot re-enter
  onHostDismissed mid-teardown.
- InAppPopoutTransport openSurface: switch host QML create to two-phase
  beginCreate/completeCreate so contentItem / backdropColor /
  dismissOnClickOutside are applied before bindings settle.
- InAppPopoutTransport: align destructor with shutdown (contentItem is
  a QObject child of hostItem, so the cascading delete handles it);
  shutdown also drops m_host/m_hostComponent so any post-shutdown
  openSurface fails the existing null guard rather than racing engine
  teardown.
- DemoController: remove dead null guard on m_controller (lifetime is
  pinned by the unique_ptr declared in the header).
- PopoutHost.qml: extract closeDuration property so the dismiss timer
  and Behaviors share a single binding source; rename QtObject id from
  _internal to dismissLatch (CLAUDE.md camelCase rule, also unblocks
  qmlformat 6.11 which fails on leading-underscore ids).

Test reshuffle (CLAUDE.md 800-line cap)
- Split test_popoutcontroller.cpp (1030 lines) into _arbitration,
  _dismiss, _signals files plus a test_popoutcontroller_helpers.h
  for the FakeTransport and ReentrantFakeTransport fixtures.
- Replace tautological destructor_detachesDismissedCallback with a
  test pinning both observable consequences via a clearCallbackCalls
  counter on the FakeTransport.
- Replace QString constData() equality (Qt-internal literal-pool
  detail) with behavioural value equality in defaultScope test.
- Replace manual new/delete QSignalSpy in destructor_with_open_popouts
  with std::unique_ptr.
- Add open_emptyIdCloseClearsEntry and
  toggle_closesDetachedEvenWhileModalActive coverage.

DRY / readability
- DemoController.toggleNamed extracts the shared toggle slot body.
- closeAll: add snapshot-before-iterate comment matching
  closeAllCooperatives; fix lambda capture comment.

CMake / packaging
- Top-level CMakeLists clang-format and qmlformat globs now cover
  libs/** and examples/**.
- libs/phosphor-popout install adds INCLUDES DESTINATION.
- PhosphorPopoutConfig.cmake.in adds Quick to find_dependency.
- Remove VERSION/SOVERSION on STATIC PhosphorPopoutQml target.
- Drop redundant set(CMAKE_AUTOMOC ON) in the demo CMakeLists.

Docs / API hygiene
- IPopoutTransport.h: document thread-affinity contract for dismissed.
- README: clarify in-tree-only QML module consumption today; document
  Component.onDestruction safety net on dismissed.
- docs/02-gap-analysis.md: rephrase ambiguous exclusive-zone wording.
- docs/04-implementation-plan.md: drop CLAUDE.md-banned for-now phrase.
- NotePopout.qml: add Accessible.description for pinned vs quick.
- main.cpp: pin Qt::DirectConnection on aboutToQuit so the synchronous
  shutdown invariant cannot regress.

Verified: cmake build green; ctest 195/195 passing including the
three split popout test binaries.
