+++ docs/phosphor-shell-design/02-gap-analysis.md
@@ -34,7 +34,7 @@ This doc is a prioritized roadmap. Each row tells you: what's missing, why it ma
 | # | Gap                                  | What it is                                                                                                | Reference impl                                                                  | Build sketch                                                                                                     | Effort | Priority |
 |---|--------------------------------------|-----------------------------------------------------------------------------------------------------------|---------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------------------------|--------|----------|
 | 1 | **Theme token singleton**            | M3 color + typography + spacing + radius + motion tokens, hot-reloaded from JSON, bound by every widget   | DMS `Common/Theme.qml` + `Appearance.qml`; Noctalia `Commons/Style.qml`         | New `qml/Phosphor/Theme/` module: `Theme.qml`, `Tokens.qml`, `Motion.qml`. Tokens fed by `IThemeService` (C++) reading a JSON file via `FileView`. Use **upstream M3 token names** verbatim. | M      | **P0**   |
-| 2 | **(shipped, PR #535) PopoutService** | One service owns lifetime / focus / screen affinity / exclusive-zone arbitration / animations for popouts | DMS `Services/PopoutService.qml` (25 KB)                                        | `phosphor-popout`: `IPopoutService`, `PopoutController`, `PopoutHost.qml`. Replaces today's `PanelPopupHost`.    | M      | ~~**P0**~~   |
+| 2 | **(shipped, PR #535) PopoutService** | One service owns lifetime / focus / screen affinity / exclusivity arbitration / animations for popouts. (Exclusivity here is the cooperative / modal / detached arbitration model, not wlr-layer-shell `exclusive_zone` routing.) | DMS `Services/PopoutService.qml` (25 KB)                                        | `phosphor-popout`: `IPopoutService`, `PopoutController`, `PopoutHost.qml`. Supersedes today's `PanelPopupHost` (migration in Phase 4.1).    | M      | ~~**P0**~~   |
 | 3 | **Widget / provider registries**     | `IBarWidgetFactory`, `IControlCenterTileFactory`, `ILauncherProviderFactory`, `IOSDFactory`, `IDesktopWidgetFactory` | Noctalia `BarWidgetRegistry` family                                  | Generalize `ILayoutSourceFactory` pattern from `phosphor-layout-api`. One header per seam; C++ register-by-name. | M      | **P0**   |
 | 4 | **Typed IPC + `phosphorctl` CLI**    | Compositor keybinds invoke shell actions by verb, not D-Bus introspection                                 | DMS `dms ipc call`, Noctalia `noctalia-shell ipc call`, QS `IpcHandler`         | `phosphor-ipc` lib: `IpcRouter`, `IpcTarget` (QML attached property), `IpcSchema` (typed). Ship `phosphorctl` Go binary. Wraps existing D-Bus adaptors. | M      | **P0**   |
 | 5 | **`PerScreen` declarative helper**   | Per-monitor surface lifecycle as a one-binding QML primitive (hotplug correctness for free)               | (Phosphor-native)                                                                | `phosphor-shell` adds `PerScreen.qml` QML helper backed by `ScreenModel`. One delegate instance per screen, lifecycle managed by `ScreenModel` add/remove signals. | S | **P0** |
