+++ docs/phosphor-shell-design/04-implementation-plan.md
@@ -262,7 +262,7 @@ The connected-corner geometry primitive, central to the visual identity.
 | `qml/Phosphor/Notifications/ToastHost.qml` + `Toast.qml`     | Layer-shell surface per screen, stacks toasts top-right. Slide-in from edge, auto-dismiss, hover-to-pause. |
 | `examples/phosphor-toast-demo/`                              | Send a notification via `notify-send` and a toast appears. Uses `phosphor-notifications` from Phase 2.5. |
 
-**Acceptance:** toasts queue, dismiss correctly; rich text + image support; respects per-app rules registered in settings (stub OK for now).
+**Acceptance:** toasts queue, dismiss correctly; rich text + image support; respects per-app rules registered in settings (stub acceptable at this phase; real rules wire in 4.3).
 
 **Effort:** M (~2 weeks)
