+++ libs/phosphor-popout/PhosphorPopoutConfig.cmake.in
@@ -5,7 +5,11 @@
 
 include(CMakeFindDependencyMacro)
 
-find_dependency(Qt6 6.6 COMPONENTS Core Gui Qml)
+# Quick is a build-time dependency for the QML module target
+# (PhosphorPopoutQml links Qt6::Quick). Consumers using only the
+# core PhosphorPopout C++ library can ignore this, but config-level
+# discovery is cheap and avoids surprises for QML consumers.
+find_dependency(Qt6 6.6 COMPONENTS Core Gui Qml Quick)
 
 include("${CMAKE_CURRENT_LIST_DIR}/PhosphorPopoutTargets.cmake")
